+++ tests/importer/test_importer.py
@@ -279,12 +279,17 @@ def test_zipimport(tmp_path, monkeypatch):
 
     zpath = tmp_path / "archive.zip"
     with ZipFile(zpath, "w") as o:
-        o.writestr("example.hy", '(setv x "Hy from ZIP")')
+        # Test a Python module as well as a Hy module to ensure that
+        # Hy's edits to ZIP imports can still get Python files.
+        o.writestr("zip_py.py", 'x = "Py from ZIP"')
+        o.writestr("zip_hy.hy", '(setv x "Hy from ZIP")')
 
     monkeypatch.syspath_prepend(zpath)
-    import example
-    assert example.x == "Hy from ZIP"
-    assert example.__file__ == str(zpath / "example.hy")
+    import zip_py, zip_hy
+    assert zip_py.x == "Py from ZIP"
+    assert zip_py.__file__ == str(zpath / "zip_py.py")
+    assert zip_hy.x == "Hy from ZIP"
+    assert zip_hy.__file__ == str(zpath / "zip_hy.hy")
 
 
 def test_eval_requiring_macro():